feat: Update routes and models for survey and survey area

+++ models/survey.py
@@ -1,5 +1,6 @@
+from sqlalchemy import ForeignKey
 from sqlalchemy.orm import relationship,Mapped, mapped_column
-from sqlalchemy.sql.sqltypes import Integer, String, ForeignKey
+from sqlalchemy.sql.sqltypes import Integer, String
 from config.db import Base
 
 # Define the table
@@ -8,10 +9,10 @@ class Survey(Base):
     id:Mapped[int] = mapped_column(Integer, primary_key=True)
     name:Mapped[str] = mapped_column(String(50), nullable=False)
     description:Mapped[str] = mapped_column(String(255), nullable=False)
-    time_range_start:Mapped[int] = mapped_column(Integer(50), nullable=False)
-    time_range_end:Mapped[int] = mapped_column(Integer(50), nullable=False)
-    time_answer_start:Mapped[int] = mapped_column(Integer(50), nullable=False)
-    time_answer_end:Mapped[int] = mapped_column(Integer(50), nullable=False)
+    time_range_start:Mapped[int] = mapped_column(Integer, nullable=False)
+    time_range_end:Mapped[int] = mapped_column(Integer, nullable=False)
+    time_answer_start:Mapped[int] = mapped_column(Integer, nullable=False)
+    time_answer_end:Mapped[int] = mapped_column(Integer, nullable=False)
     user_id:Mapped[int] = mapped_column(Integer, ForeignKey('users.id'), nullable=False)
     
     user = relationship("User", back_populates='users')
